requirements.txt
streamlit
scikit-learn
scikit-image
pillow
numpy
pandas
